requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
torch>=1.10.0
gym>=0.21.0
tqdm>=4.60.0
Pillow>=8.0.0
opencv-python>=4.5.3
scipy>=1.7.0
pandas>=1.3.0
pyserial>=3.5
keyboard>=0.13.5
stable-baselines3>=1.5.0
h5py>=3.1.0
